AURORA: Uncompressed "small" size doesn't need to be signed anymore

+++ src/aurora/smallfile.cpp
@@ -29,14 +29,14 @@
 
 namespace Aurora {
 
-static void readSmallHeader(Common::SeekableReadStream &small, int32 &type, int32 &size) {
+static void readSmallHeader(Common::SeekableReadStream &small, uint32 &type, uint32 &size) {
 	uint32 data = small.readUint32LE();
 
 	type = data & 0x000000FF;
 	size = data >> 8;
 }
 
-static void decompress00(Common::SeekableReadStream &small, Common::WriteStream &out, int32 size) {
+static void decompress00(Common::SeekableReadStream &small, Common::WriteStream &out, uint32 size) {
 	out.writeStream(small, size);
 }
 
@@ -47,13 +47,13 @@ static void decompress00(Common::SeekableReadStream &small, Common::WriteStream 
  * See <https://github.com/gravgun/dsdecmp/blob/master/CSharp/DSDecmp/Formats/Nitro/LZ10.cs#L121>
  * and <https://code.google.com/p/dsdecmp/>.
  */
-static void decompress10(Common::SeekableReadStream &small, Common::WriteStream &out, int32 size) {
+static void decompress10(Common::SeekableReadStream &small, Common::WriteStream &out, uint32 size) {
 	byte   buffer[0x10000];
 	uint32 bufferPos = 0;
 
 	uint16 flags = 0xFF00;
 
-	int32 outSize = 0;
+	uint32 outSize = 0;
 	while (outSize < size) {
 		// Only our canaries left => Read flags for the next 8 blocks
 		if (flags == 0xFF00)
@@ -74,7 +74,7 @@ static void decompress10(Common::SeekableReadStream &small, Common::WriteStream 
 
 			// Copy length bytes (and store each back into the buffer)
 			for (uint8 i = 0; i < length; i++, copyOffset++) {
-				if ((copyOffset % sizeof(buffer)) > (uint32)outSize)
+				if ((copyOffset % sizeof(buffer)) > outSize)
 					throw Common::Exception("Tried to copy past the buffer");
 
 				const byte data = buffer[copyOffset % sizeof(buffer)];
@@ -106,7 +106,7 @@ static void decompress10(Common::SeekableReadStream &small, Common::WriteStream 
 }
 
 static void decompress(Common::SeekableReadStream &small, Common::WriteStream &out,
-                       int32 type, int32 size) {
+                       uint32 type, uint32 size) {
 
 	if      (type == 0x00)
 		decompress00(small, out, size);
@@ -117,7 +117,7 @@ static void decompress(Common::SeekableReadStream &small, Common::WriteStream &o
 }
 
 void Small::decompress(Common::SeekableReadStream &small, Common::WriteStream &out) {
-	int32 type, size;
+	uint32 type, size;
 	readSmallHeader(small, type, size);
 
 	try {
@@ -130,7 +130,7 @@ void Small::decompress(Common::SeekableReadStream &small, Common::WriteStream &o
 }
 
 Common::SeekableReadStream *Small::decompress(Common::SeekableReadStream *small) {
-	int32 type, size;
+	uint32 type, size;
 	readSmallHeader(*small, type, size);
 
 	if (type == 0x00)
